Login And Registration System
Wrote a code for a login and registration system in which users can register by providing a username and password. The program checks if the username is already taken and creates a user file with the provided username and password. The user information is also stored in an in-memory data structure.